+++ src/object.rs
@@ -1,7 +1,7 @@
 pub trait Object
 {
-    fn get_position_x(&self) -> u16;
-    fn get_position_y(&self) -> u16;
+    fn get_position_x(&self) -> usize;
+    fn get_position_y(&self) -> usize;
     fn get_sprite(&self) -> &Vec<Vec<char>>;
     fn update(&self);
 }
